vbnc: Major conversion update.
Rewrite the type conversion code and implement proper conversion logic
based on the language specification instead of a pile of hacks. This also
enables the detection of a lot more type conversion errors.

Also slowly start to fix the conversion mess in CTypeExpression to not
repeat the logic elsewhere - when implicit conversions are created the
creation code now tells the CTypeExpression how to emit the conversion.