Fix Microsoft.VisualBasic.dll AssemblyVersion number
It was set to 11.0 when the net_4_5 profile was introduced in 9bae7fd107cdfff4e257282f4db5f92123916cd1.
Turns out this is actually wrong, since MS used 10.0 even for the net_4_5 version of the assembly.

This was never noticed because Mono internally remapped the assembly to the v10.0 version (see the remap logic in assembly.c),
but now with Mono 4.0 removing earlier profiles it started failing. Fixing this by setting the AssemblyVersion to the correct value.

Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=29688